Refactor turret subsystem to be shared
